Also fixed some issues in HapMixIndividualCollection: added find function for checking for masked individuals, changed type of options to HapMixOptions to avoid casting, changed name of SampleLocusAncestry to SampleHiddenStates.

Also fixed problems with manual compilation

+++ src/admixmap/HMM.cc
@@ -275,12 +275,12 @@ const pvector<double>& HMM::GetHiddenStateProbs(bool isDiploid, int t){
     States=DStates;
   }
   
-  if (hiddenStateProbs.size() != States) {
+  if (hiddenStateProbs.size() != (unsigned)States) {
     hiddenStateProbs.resize(States);
   }
 
-   const double *a = alpha + t*States;
-   const double *b = beta + t*States;  
+  const double *a = alpha + t*States;
+  const double *b = beta + t*States;  
   for( unsigned j = 0; j < States; j++ ){
     //    hiddenStateProbs[j] = alpha[t*States + j] * beta[t*States + j];
     hiddenStateProbs[j] = (*(a++)) * ( *(b++) );
